# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Management-Employees-App",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:4000/"
    }
  ],
...
  "paths": {
...
    "/users": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Created a new user",
        "description": "Created a new user",
        "operationId": "CreatedNewUser",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "exemple"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "exemple"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "exemple@email.com"
                  },
                  "phone": {
                    "type": "string",
                    "example": "phone"
                  },
                  "role": {
                    "type": "string",
                    "example": "Software Engineer"
                  },
                  "building": {
                    "type": "string",
                    "example": "Front-End"
                  },
                  "avatar": {
                    "type": "string",
                    "example": "avatar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "users"
        ],
        "summary": "Get the list of all users",
        "description": "Get the list of all users",
        "operationId": "ListUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "integer"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Field to sort by",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtBefore",
            "in": "query",
            "description": "Filter users created before this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtAfter",
            "in": "query",
            "description": "Filter users created after this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "succussffully updated ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalUsers": {
                      "type": "integer",
                      "example": 1
                    },
                    "users": {
                      "type": "array",
                      "example": [
                        {
                          "_id": "string",
                          "firstName": "string",
                          "lastName": "string",
                          "email": "exemple@email.com",
                          "phone": "string",
                          "role": "string",
                          "building": "string",
                          "avatar": "string",
                          "isActive": true,
                          "soldDays": 0,
                          "allDaysoff": 0,
                          "daysOffSick": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/daysOff": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Create a new days off ",
        "description": "Create a new days off ",
        "operationId": "AddDaysOff",
        "requestBody": {
          "description": "Created daysOff object",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDay": {
                    "type": "string",
                    "example": "2023-01-01"
                  },
                  "endDay": {
                    "type": "string",
                    "example": "2023-01-05"
                  },
                  "type": {
                    "type": "string",
                    "example": "Paid"
                  },
                  "justificationSick": {
                    "type": "string",
                    "example": "medical certificate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Days"
                }
              }
            }
          },
          "400": {
            "description": "Bad request "
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "daysoffs"
        ],
        "summary": "Get the list of all days off",
        "description": "Get the list of all days off",
        "operationId": "ListDaysOff",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "integer"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Field to sort by",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtBefore",
            "in": "query",
            "description": "Filter users created before this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          },
          {
            "name": "createdAtAfter",
            "in": "query",
            "description": "Filter users created after this date (YYYY-MM-DD format)",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "succussffully updated ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "limit": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalDaysOff": {
                      "type": "integer",
                      "example": 1
                    },
                    "daysOff": {
                      "type": "array",
                      "example": [
                        {
                          "decisionDir": {
                            "status": true,
                            "justification": "string"
                          },
                          "decisionMan": {
                            "status": false,
                            "justification": "string"
                          },
                          "_id": "string",
                          "userId": "string",
                          "startDay": "string",
                          "endDay": "string",
                          "type": "string",
                          "statusReq": false,
                          "decision": true,
                          "reqDayOff": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "request not found !"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "role",
          "building",
          "phone"
        ],
        "type": "object",
        "properties": {
          "id":{
            "type": "integer",
            "example": 10
          },
          "firstName": {
            "type": "string",
            "example": "exemple"
          },
          "lastName": {
            "type": "string",
            "example": "exemple"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "exemple@email.com"
          },
          "phone": {
            "type": "string",
            "example": "2345"
          },
          "role": {
            "type": "string",
            "example": "Software Engineer",
            "enum": [
              "Director",
              "Administration Director",
              "Administration Assistant",
              "Team manager",
              "Software Engineer",
              "Super Admin"
            ]
          },
          "building": {
            "type": "string",
            "example": "Front-End",
            "enum": [
              "Front-End",
              "Back-End",
              "Full-Stack"
            ]
          },
          "avatar": {
            "type": "string",
            "example": "12345"
          },
          "isActive": {
            "type": "boolean",
            "example": "true"
          },
          "soldDays": {
            "type": "number",
            "default": "0"
          },
          "allDaysoff": {
            "type": "number",
            "default": "0"
          },
          "daysOffSick": {
            "type": "number",
            "default": "0"
          }
        }
      },
      "Days": {
        "required": [
          "userId",
          "startDay",
          "endDay",
          "type"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "string"
          },
          "startDay": {
            "type": "string",
            "example": "2023-01-18"
          },
          "endDay": {
            "type": "string",
            "example": "2023-01-22"
          },
          "type": {
            "type": "string",
            "enum": [
              "Paid",
              "Unpaid",
              "Sick"
            ]
          },
          "decisionMan": {
            "type": "object",
            "properties": {
              "userIdMan": {
                "type": "string"
              },
              "status": {
                "type": "boolean"
              },
              "justification": {
                "type": "string"
              }
            }
          },
          "decisionDir": {
            "type": "object",
            "properties": {
              "userIdDir": {
                "type": "string"
              },
              "status": {
                "type": "boolean"
              },
              "justification": {
                "type": "string"
              }
            }
          },
          "statusReq": {
            "type": "boolean",
            "example": "true"
          },
          "decision": {
            "type": "string",
            "example": "true"
          },
          "reqDayOff": {
            "type": "number",
            "example": "0"
          },
          "justificationSick": {
            "type": "string",
            "example": "sick"
          }
        }
      }
    },
...
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}